Orchestrator/orchestrator-agent/requirements.txt
fastapi
uvicorn
python-dotenv
langchain-google-genai
langgraph
pydantic
